+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.0
 - Modified the layout
 - Added an instruction to the README
+- Changed the error messages
 
 ## 0.3.0
 - Enabled to draw cube images